Create a new branch to test the effect of my fix
Since redis does not support unit testing at function level, this branch adds some debug prints in ae.c and module.c in order to validate the fix I made and to make sure that code refactoring did not create any regression

Tests are designed to run by run.sh which will build the docker image of redis that corresponds to my fork, in the branch with the fix of integer overflow.
ae.c and module.c are copied from the aforementioned branch and were added debug prints so as to be able to check executed code.

Testing module was modified with 4 test cases. Each one of them mentions the expected results so as to be compared with the ones that are actually printed

All checks were done with file diffing between the expected and the obtained printouts